# openapi.yaml
openapi: 3.0.0
info:
  title: MarketMind AI - Financial Research Platform API
  description: >
    Production-ready REST API contract for the MarketMind AI Financial Research Platform.
    Provides endpoints for user authentication, stock listing, company profiling,
    analyst consensus tracking, stock pricing charts, earnings call transcripts,
    automated research runs, semantic vector searches (RAG), watchlists, scheduled cron jobs, and alerts.
  version: 1.0.0
servers:
  - url: https://api.marketmind.ai/v1
    description: Production API Server
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: analyst.jane@marketmind.ai
                password:
                  type: string
                  format: password
                  example: superSecretPwd123!
                first_name:
                  type: string
                  example: Jane
                last_name:
                  type: string
                  example: Doe
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Email already in use or validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Authenticate user and issue token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: analyst.jane@marketmind.ai
                password:
                  type: string
                  format: password
                  example: superSecretPwd123!
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  token_type:
                    type: string
                    example: bearer
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # -------------------------------------------------------------------------
  # STOCKS & EQUITIES
  # -------------------------------------------------------------------------
...
  /news/{symbol}:
    get:
      summary: Fetch news articles referencing a stock symbol
      tags:
        - News
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: AAPL
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: News articles successfully fetched (paginated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNewsResponse'
  /news/{symbol}/latest:
    get:
      summary: Fetch latest cached news articles from database cache
      tags:
        - News
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: AAPL
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Latest cached news articles successfully fetched (paginated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNewsResponse'
  /news/{symbol}/refresh:
    get:
      summary: Force an immediate refresh of news articles from external RSS sources
      tags:
        - News
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: AAPL
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: News articles refreshed and returned (paginated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNewsResponse'

  # -------------------------------------------------------------------------
  # RESEARCH WORKFLOW & REPORTS
  # -------------------------------------------------------------------------
...
  /alerts:
    get:
      summary: List user alerts configurations
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - name: stock_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: triggered_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Active configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
    post:
      summary: Configure a new price/sentiment trigger alert
      tags:
        - Notifications
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - alert_type
                - target_value
              properties:
                stock_id:
                  type: string
                  format: uuid
                ticker:
                  type: string
                alert_type:
                  type: string
                  enum: [price_above, price_below, sentiment_above, sentiment_below, new_research_report]
                target_value:
                  type: number
                  example: 170.0
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /alerts/{id}:
    get:
      summary: Retrieve details of a specific alert trigger
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert metadata fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
    delete:
      summary: Delete an alert trigger
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted successfully
  /alerts/evaluate/{symbol}:
    post:
      summary: Trigger immediate evaluation of active alerts for a stock symbol
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: AAPL
      responses:
        '200':
          description: Alerts evaluated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    alert_id:
                      type: string
                      format: uuid
                    alert_type:
                      type: string
                    stock_ticker:
                      type: string
                    target_value:
                      type: number
                    current_value:
                      type: number
                    triggered:
                      type: boolean
                    message:
                      type: string

  # -------------------------------------------------------------------------
  # WATCHLISTS & CRON SCHEDULES
  # -------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        created_at:
          type: string
          format: date-time
...
    NewsArticle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        summary:
          type: string
        source_name:
          type: string
        url:
          type: string
        published_at:
          type: string
          format: date-time
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        alert_type:
          type: string
        target_value:
          type: number
        is_triggered:
          type: boolean
        created_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Field validations failed
    PaginatedNewsResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NewsArticle'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...